--- src/main/java/com/t2m/g2nee/shop/bookset/Book/repository/BookCustomRepository.java
@@ -1,7 +1,0 @@
-package com.t2m.g2nee.shop.bookset.Book.repository;
-
-import org.springframework.data.repository.NoRepositoryBean;
-
-@NoRepositoryBean
-public interface BookCustomRepository {
-}
