--- src/main/java/com/t2m/g2nee/shop/bookset/Book/service/BookService.java
@@ -1,7 +1,0 @@
-package com.t2m.g2nee.shop.bookset.Book.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class BookService {
-}
