--- src/main/java/com/t2m/g2nee/shop/bookset/Contributor/domain/Contributor.java
@@ -1,22 +1,0 @@
-package com.t2m.g2nee.shop.bookset.Contributor.domain;
-
-import lombok.*;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "Contributors")
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Contributor {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long contributorId;
-    private String contributorName;
-    private String contributorEngName;
-
-}
