--- src/main/java/com/t2m/g2nee/shop/bookset/Role/domain/Role.java
@@ -1,20 +1,0 @@
-package com.t2m.g2nee.shop.bookset.Role.domain;
-
-import lombok.*;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "Roles")
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Role {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long roleId;
-    private String roleName;
-}
