--- src/main/java/com/t2m/g2nee/shop/bookset/Role/dto/RoleDto.java
@@ -1,39 +1,0 @@
-package com.t2m.g2nee.shop.bookset.Role.dto;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import javax.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-public class RoleDto {
-
-    private RoleDto() {
-    }
-
-    @Getter
-    @Setter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    @Builder
-    public static class Request {
-
-        @NotBlank(message = "역할 이름을 입력해주세요.")
-        private String roleName;
-
-    }
-
-    @Getter
-    @Setter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    @Builder
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    public static class Response {
-
-        private Long roleId;
-        private String roleName;
-    }
-}
