--- src/main/java/com/t2m/g2nee/shop/bookset/Role/repository/RoleRepository.java
@@ -1,7 +1,0 @@
-package com.t2m.g2nee.shop.bookset.Role.repository;
-
-import com.t2m.g2nee.shop.bookset.Role.domain.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoleRepository extends JpaRepository<Role,Long> {
-}
