+++ src/main/java/com/t2m/g2nee/shop/bookset/book/repository/BookCustomRepository.java
@@ -1,0 +1,7 @@
+package com.t2m.g2nee.shop.bookset.book.repository;
+
+import org.springframework.data.repository.NoRepositoryBean;
+
+@NoRepositoryBean
+public interface BookCustomRepository {
+}
