+++ src/main/java/com/t2m/g2nee/shop/bookset/book/repository/BookRepository.java
@@ -1,0 +1,7 @@
+package com.t2m.g2nee.shop.bookset.book.repository;
+
+import com.t2m.g2nee.shop.bookset.book.domain.Book;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookRepository extends JpaRepository<Book, Long> {
+}
