+++ src/main/java/com/t2m/g2nee/shop/bookset/book/service/BookService.java
@@ -1,0 +1,7 @@
+package com.t2m.g2nee.shop.bookset.book.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class BookService {
+}
