+++ src/main/java/com/t2m/g2nee/shop/bookset/role/domain/Role.java
@@ -1,0 +1,20 @@
+package com.t2m.g2nee.shop.bookset.role.domain;
+
+import lombok.*;
+
+import javax.persistence.*;
+
+@Entity
+@Table(name = "Roles")
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Role {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long roleId;
+    private String roleName;
+}
