+++ src/main/java/com/t2m/g2nee/shop/bookset/role/repository/RoleRepository.java
@@ -1,0 +1,7 @@
+package com.t2m.g2nee.shop.bookset.role.repository;
+
+import com.t2m.g2nee.shop.bookset.role.domain.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RoleRepository extends JpaRepository<Role,Long> {
+}
