+++ src/test/java/com/t2m/g2nee/shop/bookset/publisher/controller/PublisherControllerTest.java
@@ -30,7 +30,8 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 
-@WebMvcTest(PublisherController.class)
+@SpringBootTest
+@AutoConfigureMockMvc
 class PublisherControllerTest {
 
     @Autowired
