+++ src/test/java/com/t2m/g2nee/shop/bookset/tag/controller/TagControllerTest.java
@@ -26,10 +26,12 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
+import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 
-@WebMvcTest(TagController.class)
+@SpringBootTest
+@AutoConfigureMockMvc
 class TagControllerTest {
     @Autowired
     private MockMvc mockMvc;
